requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
torch==2.0.1
tensorflow==2.13.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.2
geopy==2.3.0
python-dotenv==1.0.0
requests==2.31.0
PyJWT==2.8.0
celery==5.3.1
redis==4.6.0
psycopg2-binary==2.9.7
gunicorn==21.2.0
prometheus-client==0.17.1
APScheduler==3.10.4